combined sdfClient and suiteAppClient to a single client
