+++ packages/netsuite-adapter/src/adapter.ts
@@ -20,7 +20,6 @@ import {
 import _ from 'lodash'
 import { collections, values } from '@salto-io/lowerdash'
 import { resolveValues } from '@salto-io/adapter-utils'
-import NetsuiteClient from './client/client'
 import {
   createInstanceElement, getLookUpName, toCustomizationInfo,
 } from './transformer'
@@ -41,6 +40,7 @@ import {
 import { getAllReferencedInstances, getRequiredReferencedInstances } from './reference_dependencies'
 import { andQuery, buildNetsuiteQuery, NetsuiteQueryParameters, notQuery } from './query'
 import { SuiteAppClient } from './client/suiteapp_client/suiteapp_client'
+import { NetsuiteClient } from './client/client'
 
 const { makeArray } = collections.array
