+++ packages/netsuite-adapter/test/changes_detector/changes_detector.test.ts
@@ -20,6 +20,8 @@ import * as fileCabinetDetector from '../../src/changes_detector/changes_detecto
 import { customRecordTypeDetector } from '../../src/changes_detector/changes_detectors/custom_type'
 import scriptDetector from '../../src/changes_detector/changes_detectors/script'
 import { getChangedObjects } from '../../src/changes_detector/changes_detector'
+import { NetsuiteClient } from '../../src/client/client'
+import mockSdfClient from '../client/sdf_client'
 
 describe('changes_detector', () => {
   const query = {
@@ -31,10 +33,11 @@ describe('changes_detector', () => {
   const getChangesFoldersMock = jest.spyOn(fileCabinetDetector, 'getChangedFolders').mockResolvedValue([])
 
   const runSavedSearchQueryMock = jest.fn()
-  const client = {
+  const suiteAppClient = {
     runSavedSearchQuery: runSavedSearchQueryMock,
   } as unknown as SuiteAppClient
 
+  const client = new NetsuiteClient(mockSdfClient(), suiteAppClient)
   beforeEach(() => {
     jest.resetAllMocks()
     getCustomRecordTypeChangesMock.mockResolvedValue([
