+++ packages/netsuite-adapter/test/changes_detector/custom_list.test.ts
@@ -16,13 +16,16 @@
 import { SuiteAppClient } from '../../src/client/suiteapp_client/suiteapp_client'
 import { customListDetector as detector } from '../../src/changes_detector/changes_detectors/custom_type'
 import { Change } from '../../src/changes_detector/types'
+import mockSdfClient from '../client/sdf_client'
+import { NetsuiteClient } from '../../src/client/client'
 
 describe('custom_list', () => {
   const runSuiteQLMock = jest.fn()
-  const client = {
+  const suiteAppClient = {
     runSuiteQL: runSuiteQLMock,
   } as unknown as SuiteAppClient
 
+  const client = new NetsuiteClient(mockSdfClient(), suiteAppClient)
 
   describe('query success', () => {
     let results: Change[]
