+++ packages/netsuite-adapter/test/changes_detector/file_cabinet.test.ts
@@ -16,13 +16,16 @@
 import { SuiteAppClient } from '../../src/client/suiteapp_client/suiteapp_client'
 import { getChangedFiles, getChangedFolders } from '../../src/changes_detector/changes_detectors/file_cabinet'
 import { Change } from '../../src/changes_detector/types'
+import mockSdfClient from '../client/sdf_client'
+import { NetsuiteClient } from '../../src/client/client'
 
 describe('file_cabinet', () => {
   const runSuiteQLMock = jest.fn()
-  const client = {
+  const suiteAppClient = {
     runSuiteQL: runSuiteQLMock,
   } as unknown as SuiteAppClient
 
+  const client = new NetsuiteClient(mockSdfClient(), suiteAppClient)
   describe('getChangedFiles', () => {
     describe('query success', () => {
       let results: Change[]
