+++ packages/netsuite-adapter/test/changes_detector/script.test.ts
@@ -16,12 +16,16 @@
 import { SuiteAppClient } from '../../src/client/suiteapp_client/suiteapp_client'
 import detector, { SCRIPT_TYPES } from '../../src/changes_detector/changes_detectors/script'
 import { Change } from '../../src/changes_detector/types'
+import { NetsuiteClient } from '../../src/client/client'
+import mockSdfClient from '../client/sdf_client'
 
 describe('script', () => {
   const runSuiteQLMock = jest.fn()
-  const client = {
+  const suiteAppClient = {
     runSuiteQL: runSuiteQLMock,
   } as unknown as SuiteAppClient
+
+  const client = new NetsuiteClient(mockSdfClient(), suiteAppClient)
 
   describe('script fields changes', () => {
     let results: Change[]
